+++ riscv/src/register.rs
@@ -89,8 +89,12 @@ pub mod mepc;
 pub mod mip;
 pub mod mscratch;
 pub mod mtinst;
+pub mod mtopi;
 pub mod mtval;
 pub mod mtval2;
+pub mod mvien;
+#[cfg(any(test, target_arch = "riscv32"))]
+pub mod mvienh;
 
 // Machine Protection and Translation
 mod pmpcfgx;