+++ riscv/src/register.rs
@@ -132,3 +132,5 @@ mod tests;
 // TODO: Debug Mode Registers
 pub mod dcsr;
 pub mod dpc;
+pub mod dscratch0;
+pub mod dscratch1;