+++ riscv/src/register/sip.rs
@@ -1,7 +1,6 @@
 //! sip register
 
-use crate::bits::bf_extract;
-use riscv_pac::CoreInterruptNumber;
+use crate::{bits::bf_extract, CoreInterruptNumber};
 
 read_only_csr! {
     /// sip register