+++ riscv/CHANGELOG.md
@@ -16,8 +16,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - New `rt` and `rt-v-trap` features to opt-in `riscv-rt`-related code in `riscv::pac_enum` macro.
 - Add `mvien` + `mvienh` CSR
 
-# Changed
+### Changed
 
+- Moved macros from `./macros/` to `../riscv-macros/`
 - Updated the license to `MIT or Apache-2.0`
 - Bump MSRV to 1.68 for latest version of syn 2.0
 - Now, `riscv::pac_enum` macro only includes trap-related code if `rt` or `rt-v-trap` features are enabled.